chore: Extracted new dataclass into separate file
